+++ tests/xy_c_compile_resources/callbacks/callbacks2.c
@@ -19,7 +19,7 @@ void callbacks2_test2(void) {
 
 void callbacks2_invokeTests(void) {
     const xy_fp__void l_tests[3] = {callbacks2_test1, callbacks2_test2, callbacks2_test3};
-    for (size_t tmp_0_iter = 0; tmp_0_iter < 3; ++tmp_0_iter) {
+    for (uint64_t tmp_0_iter = 0; tmp_0_iter < 3; ++tmp_0_iter) {
         l_tests[tmp_0_iter]();
     }
 }
