+++ tests/xy_c_compile_resources/tags/positionalTags.c
@@ -6,14 +6,14 @@ typedef struct positionalTags_Array positionalTags_Array;
 typedef struct positionalTags_Point positionalTags_Point;
 
 struct positionalTags_Array {
-    size_t m_len;
+    uint64_t m_len;
 };
 struct positionalTags_Point {
     float m_x;
     float m_y;
 };
 
-size_t positionalTags_test(void) {
+uint64_t positionalTags_test(void) {
     positionalTags_Array l_ints = {0};
     positionalTags_Array l_points = {0};
     return l_ints.m_len + l_points.m_len + 4 + 8;
